fix build migrate to Go modules

+++ Makefile
@@ -1,21 +1,22 @@
 BUILD_DIR  ?= build
 GOX_ARCH   ?= darwin/386 darwin/amd64 linux/amd64 linux/386
 GOX_OUTPUT ?= $(BUILD_DIR)/{{.Dir}}_{{.OS}}_{{.Arch}}
+GOMODULE   ?= GO111MODULE=on
 
 default: install test build
 
 install:
-	go get -t -v ./...
-	go get github.com/mitchellh/gox
+	$(GOMODULE) go get -t -v ./...
+	$(GOMODULE) go get github.com/mitchellh/gox
 
 test:
-	go test ./...
+	$(GOMODULE) go test ./...
 
 test.coverage:
-	go test ./... -coverprofile=coverage.txt -covermode=atomic
+	$(GOMODULE) go test ./... -coverprofile=coverage.txt -covermode=atomic
 
 build: build.clean
-	gox -output "$(GOX_OUTPUT)" -osarch="$(GOX_ARCH)"
+	$(GOMODULE) gox -output "$(GOX_OUTPUT)" -osarch="$(GOX_ARCH)"
 
 build.clean:
-	rm $(BUILD_DIR)/*
+	if [ -d $(BUILD_DIR) ]; then rm $(BUILD_DIR)/* ; fi
